% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_credit_plots.R
\name{assign_pc}
\alias{assign_pc}
\title{Perform partial credit analysis given a grade key}
\usage{
assign_pc(door, pcvalues, n1, n2)
}
\arguments{
\item{door}{a data frame of DOOR outcome distribution}

\item{pcvalues}{a numeric vector of grade key}

\item{n1, n2}{samples size for group 1, 2}
}
\value{
results from a partial credit analysis
}
\description{
Perform partial credit analysis given a grade key
}
\keyword{internal}
