% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/door_barplot.R
\name{door_barplot}
\alias{door_barplot}
\title{Create DOOR summary barplot}
\usage{
door_barplot(
  y1 = NULL,
  y2 = NULL,
  summary_obj = NULL,
  data_type = c("freq", "prop")
)
}
\arguments{
\item{y1, y2}{Numeric vectors of DOOR proportion or frequency distribution for
group 1, group 2.
The entries should be ordered from most desirable to least desirable}

\item{summary_obj}{An object returned by \code{door_summary()}; Alternative
input for y1 and y2}

\item{data_type}{Either "freq" for frequency input or "prop" for proportion input
when using y1 and y2}
}
\value{
a ggplot object
}
\description{
Create DOOR summary barplot
}
\examples{
y1 = c(60, 30, 10)
y2 = c(50, 40, 10)
door_barplot(y1, y2)
}
