% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/door_barplot.R
\name{generate_color}
\alias{generate_color}
\title{Generate DOOR standard color scheme}
\usage{
generate_color(n)
}
\arguments{
\item{n}{Number of DOOR categories}
}
\value{
colors
}
\description{
Generate DOOR standard color scheme
}
\keyword{internal}
