% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{generate_forestp}
\alias{generate_forestp}
\title{Forest plot render}
\usage{
generate_forestp(data, text, xlab, issum, xticks, line_height = 1.3)
}
\arguments{
\item{data}{a data frame for numerical values of the forest plot}

\item{text}{a data frame for the text part of the forest plot}

\item{xlab}{a character string of x-axis label}

\item{issum}{a logical vector indicating the summary row}

\item{xticks}{a vector specifying the x-axis tick marks}

\item{line_height}{a numeric value specifying the space between the rows}
}
\value{
a forest plot object
}
\description{
Forest plot render
}
\keyword{internal}
