% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_credit_plots.R
\name{generate_shape}
\alias{generate_shape}
\title{Generate shapes for partial credit plot}
\usage{
generate_shape(n)
}
\arguments{
\item{n}{number of shapes to generate}
}
\value{
a vector of code shapes
}
\description{
Generate shapes for partial credit plot
}
\keyword{internal}
