% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_door_probability.R
\name{get_door_probability}
\alias{get_door_probability}
\title{Get DOOR Probability}
\usage{
get_door_probability(res, tx = c("A", "B"))
}
\arguments{
\item{res}{summary table of the sample size by treatment and door outcome, result of get_door_summary. Treatment in columns, DOOR outcomes in rows.}

\item{tx}{character, levels of the treatment indicator variable that should also be column names res}
}
\value{
dataframe with the door summary
}
\description{
This function calculates the DOOR probability from the door summary.
}
\examples{
set.seed(123)
N <- 26
seq <- rep(c("A", "B"), each = N/2)
DOOR <- sample(1:4, size = N, replace = TRUE)
data <- data.frame(seq, DOOR)
res <- get_door_summary(data, "seq", "DOOR")
get_door_probability(res)


}
\references{
Evans, S. R., Rubin, D., Follmann, et. al. (2015). Desirability of outcome ranking (DOOR) and response adjusted for duration of antibiotic risk (RADAR). Clinical Infectious Diseases, 61(5), 800-806.
}
