% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Install statgarten dependencies.}
\usage{
init()
}
\value{
an HTML list
}
\description{
Install statgarten dependencies, datatoys, board, scissor, colorpen, soroban, stove
}
\examples{
init()
}
