% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/door_ci.R
\name{inv.logit}
\alias{inv.logit}
\title{Calculate inverse of logit transformation}
\usage{
inv.logit(x)
}
\arguments{
\item{x}{A numerical value}
}
\value{
Back transformed value
}
\description{
Calculate inverse of logit transformation
}
\keyword{internal}
