% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_raw_data}
\alias{mock_raw_data}
\title{Mock Raw Data}
\format{
A data frame with 55 observations with the following columns
\describe{
\item{ARM, Arm text}{Codes and labels of treatment arm}
\item{DOOR, DOORtext}{Codes and labels of DOOR outcome}
\item{clinical failure, clinical failure text}{Codes and labels for one of the DOOR components}
\item{infectious complications, infectious complications text}{Codes and labels for one of the DOOR components}
\item{death, death text}{Codes and labels for one of the DOOR components}
\item{Weight}{IPW weights}
\item{Duration}{Tie breaker}
}
}
\usage{
data(mock_raw_data)
}
\description{
A mock data that contains examples of a raw dataset of DOOR outcomes, treatment
information, and DOOR components
}
\keyword{datasets}
