% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_credit_plots.R
\name{partial_credit_biplot}
\alias{partial_credit_biplot}
\title{Partial credit plot}
\usage{
partial_credit_biplot(pc_object, ...)
}
\arguments{
\item{pc_object}{an object returned by \code{partial_credit_analysis()}}

\item{...}{additional arguments for other functions}
}
\value{
a plot object
}
\description{
Partial credit plot
}
\examples{
grade.key <- c(100, 60, 0)
y1 <- c(60, 30, 10)
y2 <- c(50, 40, 10)
pc_object <- partial_credit_analysis(grade_key = grade.key, y1 = y1, y2 = y2)
partial_credit_biplot(pc_object)
}
