% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_credit_plots.R
\name{pc.contour}
\alias{pc.contour}
\title{Calculates values for contour plots}
\usage{
pc.contour(DOOR, n_inc = 1)
}
\arguments{
\item{DOOR}{a data frame of DOOR outcome distribution}

\item{n_inc}{size of increments of partial credits}
}
\value{
calculated values for the contour plot
}
\description{
Calculates values for contour plots
}
\keyword{internal}
