% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DOOR_comparison.R
\name{plot_door_comparison}
\alias{plot_door_comparison}
\title{Plot DOOR comparison}
\usage{
plot_door_comparison(res)
}
\arguments{
\item{res}{door summary}
}
\value{
barplot with the DOOR comparison
}
\description{
This function plots the DOOR comparison
}
\examples{
set.seed(123)
N <- 26
seq <- rep(c("A", "B"), each = N/2)
DOOR <- sample(1:4, size = N, replace = TRUE)
data <- data.frame(seq, DOOR)
res <- get_door_summary(data, "seq", "DOOR")

}
\references{
Evans, S. R., Rubin, D., Follmann, et. al. (2015). Desirability of outcome ranking (DOOR) and response adjusted for duration of antibiotic risk (RADAR). Clinical Infectious Diseases, 61(5), 800-806.
}
