% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/door_ci.R
\name{reml}
\alias{reml}
\title{Restricted maximum likelihood estimate of DOOR outcome cell proportions}
\usage{
reml(y1, y2, theta.null)
}
\arguments{
\item{y1, y2}{vector of frequency distribution of DOOR outcome for group 1, 2}

\item{theta.null}{The DOOR probability of interest}
}
\value{
An object for REML estimates
}
\description{
Restricted maximum likelihood estimate of DOOR outcome cell proportions
}
\keyword{internal}
