% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_N300}
\alias{sample_data_N300}
\title{DOOR Sample Data (N = 300)}
\format{
A data frame with 300 rows and 2 variables:
\describe{
  \item{seq}{treament assignment sequence}
  \item{DOOR}{DOOR outcome with four categories}
}
}
\usage{
sample_data_N300
}
\description{
An artificial data set containing the the treatment assignment
vector and the DOOR outcome.
}
\keyword{datasets}
