% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_credit_analysis.R
\name{t_test_summary}
\alias{t_test_summary}
\title{t test with summary statistics}
\usage{
t_test_summary(
  m1,
  m2,
  s1,
  s2,
  n1,
  n2,
  m0 = 0,
  equal.variance = FALSE,
  conf = 0.95
)
}
\arguments{
\item{m1, m2}{means of each group}

\item{s1, s2}{standard deviation of each group}

\item{n1}{sample size of each group}

\item{m0}{value of null hypothesis}

\item{equal.variance}{whether to assume equal variance between the groups}

\item{conf}{confidence level for confidence interval}
}
\value{
an "htest" object of the test result
}
\description{
t test with summary statistics
}
\keyword{internal}
