% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_pi.R
\name{var_pi}
\alias{var_pi}
\title{Variance of DOOR probability}
\usage{
var_pi(p1, p2, n1, n2)
}
\arguments{
\item{p1}{Vector of DOOR outcome proportion distribution for group 1}

\item{p2}{Vector of DOOR outcome proportion distribution for group 2}

\item{n1}{Sample size of group 1}

\item{n2}{Sample size of group 2}
}
\value{
DOOR probability, Wald-type variance, exact variance,
Halperin variance; theta for Halperin method
}
\description{
Calculate multiple types of variances of DOOR probability
}
